pyproject.toml
[project]
name = "db-query-assistant"
version = "0.1.0"
description = ""
authors = [
    {name = "ZhaoYT",email = "yatong.zhao@foxmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pydantic-ai-slim[openai] (>=0.2.16,<0.3.0)",
    "aiomysql (>=0.2.0,<0.3.0)",
    "pyecharts (>=2.0.8,<3.0.0)",
]

[tool.poetry]
packages = [{include = "db_query_assistant", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
